requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
discord.py==1.4.1
docopt==0.6.2
idna==2.9
isort==4.3.21
Js2Py==0.70
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
multidict==4.7.6
packaging==20.4
Pillow==7.1.2
pipwin==0.5.0
praw==7.1.0
prawcore==1.5.0
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pycparser==2.20
PyGetWindow==0.0.8
pyjokes==0.6.0
pyjsparser==2.7.1
pylint==2.5.2
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMsgBox==1.0.8
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.0
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.3
python-dotenv==0.14.0
pyttsx3==2.88
PyTweening==1.0.3
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
toml==0.10.1
tzlocal==2.1
update-checker==0.18.0
urllib3==1.25.9
websocket-client==0.57.0
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.11.1.1
